Modified builder/diagnostics to call raster_utils function to smooth the QA mask differently
QA mask can now be filtered with a maximum filter to create a buffer around any mask pixel, or convolve the mask with a kernel to better smooth out odd pixels in the mask that are usually false positives.
Also moved some variables relating to masking from config.py back to builder.main() as arguments - better to control in call than hardcoded.

 On branch smooth_qamask
 Changes to be committed:
	modified:   process/builder.py
	modified:   process/common/raster_utils.py
	modified:   process/config.py
	modified:   process/diagnostics.py