Changed way of defining threshold for convolved mask
Instead of setting threshold to width/sqrt(2), now set it to area/8.
A bit of experimentation suggested the factor 8.  This is more
meaningful than the previous method, now it checks whether the
number of masked pixels within the convolution kernel is greater
than 1/8 of all pixels in the kernel. However, the next thing is
to add this factor to the config or parameters, rather than it
being hardcoded.

 On branch circular_qamask
 Your branch is up to date with 'origin/circular_qamask'.

 Changes to be committed:
	modified:   process/builder.py
	modified:   process/common/raster_utils.py
	modified:   process/diagnostics.py